requirements.txt
anyio==3.6.1
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.6.15
cycler==0.11.0
emoji==2.0.0
fonttools==4.34.4
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
isOdd==0.1.2
kiwisolver==1.4.4
matplotlib==3.5.2
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
progress==1.6
psycopg2==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
tornado==6.2
tzdata==2022.1
tzlocal==4.2
